Cambie muhco el unitmanager, hice los prefbas de ataques, cambie codigo y como s eponen los tiles, cree un prefab de tile unico que debemos usar para la grid
