+++ Assets/Managers/GridManager.cs
@@ -7,7 +7,7 @@ public class GridManager : MonoBehaviour
 {
     [SerializeField] public int _width, _height;
 
-    [SerializeField] private Tile _grassfile, _enemytile;
+    [SerializeField] private Tile tile;
 
     [SerializeField] private Transform _cam;
 
@@ -32,13 +32,26 @@ public class GridManager : MonoBehaviour
         {
             for (int y = 0; y < _height; y++)
             {
-                var typeOfTile = x < _width / 2 ? _grassfile : _enemytile;
-                var spawnedTile = Instantiate(typeOfTile, new Vector3(x, y), Quaternion.Euler(90, 0, 0));
+                var typeoftile = tile;
+                if(x < _width / 2)
+                {
+                    typeoftile.EnemyTile.gameObject.SetActive(false);
+                    typeoftile.HeroTile.gameObject.SetActive(true);
+                    typeoftile.Faction = Faction.Hero;
+
+                }else if(x >= _width / 2)
+                {
+                    typeoftile.HeroTile.gameObject.SetActive(false);
+                    typeoftile.EnemyTile.gameObject.SetActive(true);
+                    typeoftile.Faction = Faction.Enemy;
+                }
+                //var typeOfTile = x < _width / 2 ? _grassfile : _enemytile;
+                var spawnedTile = Instantiate(typeoftile, new Vector3(x, y), Quaternion.Euler(90, 0, 0));
                 spawnedTile.name = $"Tile {x} {y}";
 
                 var isOffset = (x % 2 == 0 && y % 2 != 0) || (x % 2 != 0 && y % 2 == 0);
-                spawnedTile.Init(isOffset);
-
+                spawnedTile.HeroTile.Init(isOffset);
+                spawnedTile.EnemyTile.Init(isOffset);
 
                 _tiles[new Vector2(x, y)] = spawnedTile;
                 spawnedTile.x = x;
