+++ Assets/Prefabs/Units/Attacks/DruidAttacks/AtaqueDruida (3).prefab
@@ -125,15 +125,19 @@ MonoBehaviour:
   UnitName: AeO efecto
   OccupiedTile: {fileID: 0}
   Damage: 5
+  Heal: 0
   Faction: 0
   AreaOfEffect: 2
   CoolDown: 5
   LastCast1: 0
   LastCast2: -100
   ManaCost: 0.5
-  DoVeneno: 5
+  invocacion: {fileID: 0}
+  DoVeneno: 10
+  DoBurn: 0
+  stun: 0
   attackType: 0
-  type: 0
+  type: 2
 --- !u!1 &3037364760383503627
 GameObject:
   m_ObjectHideFlags: 0
