+++ Assets/Prefabs/Units/Attacks/PlantsAttacks/AtaqueQuemar.prefab
@@ -124,12 +124,12 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   UnitName: EstadoEnemigos
   OccupiedTile: {fileID: 0}
-  Damage: 0
+  Damage: 1
   Heal: 0
   Faction: 0
   AreaOfEffect: 1
   CoolDown: 15
-  LastCast1: 0
+  LastCast1: 10
   LastCast2: -100
   ManaCost: 0
   invocacion: {fileID: 0}
