+++ Assets/Prefabs/Units/Attacks/RobotAttacks/AtaqueRobot3.prefab
@@ -134,7 +134,7 @@ MonoBehaviour:
   ManaCost: 5
   invocacion: {fileID: 7735570308767223746, guid: fdd801f139e7d7b47a848087e5576ac2,
     type: 3}
-  DoVeneno: 10
+  DoVeneno: 0
   DoBurn: 0
   stun: 0
   attackType: 0
