+++ Assets/Prefabs/Units/Heroes/DruidaImpsotor.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2e475f0b150a10f45a8a20fbff88c12a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
